ZOOKEEPER-3970: Enable ZooKeeperServerController to expire session. (#147)
This is a follow up of ZOOKEEPER-3948. Here we enable ZooKeeperServerController to be able to expire a global or local session. This is very useful in our experience in integration testing when we want a controlled session expiration mechanism. This is done by having session tracker exposing both global and local session stats, so a zookeeper server can expire the sessions in the controller.

Author: Michael Han <hanm@apache.org>

Reviewers: Damien Diederen <ddiederen@apache.org>

Closes #1505 from hanm/ZOOKEEPER-3970

Co-authored-by: Michael Han <hanm@apache.org>