ZOOKEEPER-4312: ZooKeeperServerEmbedded: enhance server start/stop for testability (#153)
see https://issues.apache.org/jira/browse/ZOOKEEPER-4312

ZooKeeperServerEmbedded works well for running ZooKeeper but it lacks support for a few little features in order to use it for tests.

I saw these problems while working on the port of Curator Testing Server to ZooKeeperServerEmbedded.

There is no wait to wait for the server to be up-and-running
When you "close()" the server, it does not wait for the ports to be closed
There is no wait to have the ConnectString for the server

Author: Enrico Olivelli <eolivelli@apache.org>

Reviewers: Damien Diederen <ddiederen@apache.org>

Closes #1710 from eolivelli/impl/embedded-2

Co-authored-by: Enrico Olivelli <eolivelli@apache.org>